+++ packages/chat/components/feedback/src/feedback.scss
@@ -11,10 +11,10 @@ $css--plex: true !default;
 @use '@carbon/styles/scss/themes' as *;
 @use '@carbon/web-components/scss/components/button/button';
 
-:host(#{$c4ai-prefix}-feedback-container) {
+:host(#{$c4ai-prefix}-feedback) {
   position: relative;
 }
 
-:host(#{$c4ai-prefix}-feedback-container) .feedback-flag {
+:host(#{$c4ai-prefix}-feedback) .feedback-flag {
   position: absolute;
 }
