+++ packages/chat/components/test-input/__stories__/test-input.stories.js
@@ -7,7 +7,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import '../components/test-input/test-input';
+import '../test-input';
 import { html } from 'lit';
 
 // More on how to set up stories at: https://storybook.js.org/docs/web-components/writing-stories/introduction
